Missing int to long and start updating release notes

+++ src/OrchardCore.Modules/OrchardCore.Search.Lucene/Services/LuceneIndexingService.cs
@@ -55,7 +55,7 @@ namespace OrchardCore.Search.Lucene
         public async Task ProcessContentItemsAsync(string indexName = default)
         {
             // TODO: Lock over the filesystem in case two instances get a command to rebuild the index concurrently.
-            var allIndices = new Dictionary<string, int>();
+            var allIndices = new Dictionary<string, long>();
             var lastTaskId = Int64.MaxValue;
             IEnumerable<LuceneIndexSettings> indexSettingsList = null;
